Add TemplateEditor components for email and in-app editing
- Introduced EmailEditor and InAppEditor components to handle email and in-app template editing.
- Added LANGUAGE_CODES constant for language selection in editors.
- Updated TemplateEditor to support new editor components based on template type.
- Removed legacy EmailEditor implementation from the email directory.
