+++ ui-app/client/src/components/TemplateEditor/editors/EmailEditorIcons.tsx
